requirements.txt
# core LLM + LangChain stack
langchain>=0.2.17
langchain-community>=0.2.17

# Google AI integration
langchain-google-genai>=2.1.8
google-generativeai>=0.8.5

# OpenAI (for embeddings only - Gemini doesn't provide embeddings)
openai>=1.84.0

# vector store
faiss-cpu>=1.8.0        # CPU-only FAISS binaries

# document loading / misc
pypdf>=4.0.0            # used by langchain_community.document_loaders.PyPDFLoader
python-dotenv>=1.0.1    # load API keys from .env
tiktoken>=0.6.0         # fast token counting (speeds up splitters)

# optional - add later if you OCR scanned PDFs
# unstructured[ocr]>=0.12.0
# pillow>=10.0.0
